requirements-docker-cuda124.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
numpy==2.2.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
six==1.17.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
